refactor: separated pydantic IF types and instantiation models  from the config
