+++ src/modalities/inference/text/config.py
@@ -1,33 +1,13 @@
-from typing import Annotated, Dict, Optional
+from typing import Optional
 
-from pydantic import BaseModel, FilePath, field_validator
+from pydantic import BaseModel, field_validator
 
-from modalities.config.config import (
+from modalities.config.pydanctic_if_types import (
     PydanticPytorchDeviceType,
     PydanticPytorchModuleType,
-    PydanticThirdPartyTypeIF,
     PydanticTokenizerIFType,
 )
 from modalities.config.utils import parse_torch_device
-from modalities.inference.text.inference_component import TextInferenceComponent
-
-PydanticTextInferenceComponentType = Annotated[TextInferenceComponent, PydanticThirdPartyTypeIF(TextInferenceComponent)]
-
-
-class TextGenerationSettings(BaseModel):
-    model_path: FilePath
-    context_length: int
-    device: PydanticPytorchDeviceType
-    referencing_keys: Dict[str, str]
-
-    @field_validator("device", mode="before")
-    def parse_device(cls, device) -> PydanticPytorchDeviceType:
-        return parse_torch_device(device)
-
-
-class TextGenerationInstantiationModel(BaseModel):
-    text_inference_component: PydanticTextInferenceComponentType
-    settings: TextGenerationSettings
 
 
 class TextInferenceComponentConfig(BaseModel):
